feat: esm with type module
