+++ docs/posts/check-package.md
@@ -1,0 +1,10 @@
+---
+title: "Linting npm packages"
+date: "2023-02-23"
+---
+
+Lint npm packages using these sites:
+
+- https://publint.dev/
+- https://arethetypeswrong.github.io/
+- https://bundlephobia.com/
