+++ package.json
@@ -2,6 +2,7 @@
   "name": "blog",
   "version": "0.0.7",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "next dev",
     "build": "next build",
@@ -9,7 +10,7 @@
     "change:add": "changeset add"
   },
   "dependencies": {
-    "next": "^13.1.6",
+    "next": "^13.2.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
@@ -26,7 +27,7 @@
     "autoprefixer": "^10.4.13",
     "date-fns": "^2.16.1",
     "gray-matter": "^4.0.2",
-    "hazel-ui": "^0.4.0",
+    "hazel-ui": "^0.4.2",
     "husky": "^8.0.2",
     "next-mdx-remote": "^4.2.1",
     "postcss": "^8.4.20",
@@ -36,7 +37,7 @@
     "rehype-slug": "^4.0.1",
     "remark-code-titles": "^0.1.1",
     "tailwindcss": "^3.2.4",
-    "typescript": "^4.9.4",
+    "typescript": "^5.0.0-beta",
     "windy-radix-palette": "^0.4.1"
   },
   "nextBundleAnalysis": {
