--- src/Home/DefaultTags.tsx
@@ -1,11 +1,0 @@
-import { config } from "../config";
-
-export default function DefaultTags() {
-  return (
-    <>
-      <link rel="icon" href={`${config.basePath}/favicon.ico`} />
-      <meta name="twitter:card" content="summary_large_image" />
-      <meta name="twitter:creator" content="@cse_as" />
-    </>
-  );
-}
