--- src/Home/components/index.ts
@@ -1,2 +1,0 @@
-export { About } from "./About";
-export { RecentPosts } from "./RecentPosts";
